+++ glean-core/android/src/test/java/mozilla/telemetry/glean/private/LabeledMetricTypeTest.kt
@@ -397,7 +397,7 @@ class LabeledMetricTypeTest {
 
     @Test(expected = IllegalStateException::class)
     fun `Test that labeled events are an exception`() {
-        val eventMetric = EventMetricType<NoExtraKeys>(
+        val eventMetric = EventMetricType<NoReasonCodes>(
             disabled = false,
             category = "telemetry",
             lifetime = Lifetime.Application,
@@ -405,7 +405,7 @@ class LabeledMetricTypeTest {
             sendInPings = listOf("metrics")
         )
 
-        val labeledEventMetric = LabeledMetricType<EventMetricType<NoExtraKeys>>(
+        val labeledEventMetric = LabeledMetricType<EventMetricType<NoReasonCodes>>(
             disabled = false,
             category = "telemetry",
             lifetime = Lifetime.Application,