+++ glean-core/android/src/test/java/mozilla/telemetry/glean/scheduler/MetricsPingSchedulerTest.kt
@@ -10,9 +10,11 @@ import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.work.testing.WorkManagerTestInitHelper
 import mozilla.components.support.test.any
+import mozilla.components.support.test.eq
 import mozilla.telemetry.glean.Dispatchers
 import mozilla.telemetry.glean.getContextWithMockedInfo
 import mozilla.telemetry.glean.Glean
+import mozilla.telemetry.glean.GleanMetrics.Pings
 import mozilla.telemetry.glean.private.Lifetime
 import mozilla.telemetry.glean.resetGlean
 import mozilla.telemetry.glean.private.StringMetricType
@@ -36,7 +38,6 @@ import org.junit.runner.RunWith
 import org.mockito.Mockito.anyBoolean
 import org.mockito.Mockito.anyString
 import org.mockito.Mockito.doReturn
-import org.mockito.Mockito.eq
 import org.mockito.Mockito.never
 import org.mockito.Mockito.spy
 import org.mockito.Mockito.times
@@ -235,16 +236,18 @@ class MetricsPingSchedulerTest {
         verify(mpsSpy, times(0)).updateSentDate(anyString())
         verify(mpsSpy, times(0)).schedulePingCollection(
             any(),
-            anyBoolean()
+            anyBoolean(),
+            eq(Pings.metricsReasonCodes.overdue)
         )
 
-        mpsSpy.collectPingAndReschedule(Calendar.getInstance())
+        mpsSpy.collectPingAndReschedule(Calendar.getInstance(), false, Pings.metricsReasonCodes.overdue)
 
         // Verify that we correctly called in the methods.
         verify(mpsSpy, times(1)).updateSentDate(anyString())
         verify(mpsSpy, times(1)).schedulePingCollection(
             any(),
-            anyBoolean()
+            anyBoolean(),
+            eq(Pings.metricsReasonCodes.reschedule)
         )
     }
 
@@ -274,7 +277,11 @@ class MetricsPingSchedulerTest {
             assertTrue("The initial test data must have been recorded", testMetric.testHasValue())
 
             // Manually call the function to trigger the collection.
-            Glean.metricsPingScheduler.collectPingAndReschedule(Calendar.getInstance())
+            Glean.metricsPingScheduler.collectPingAndReschedule(
+                Calendar.getInstance(),
+                false,
+                Pings.metricsReasonCodes.overdue
+            )
 
             // Trigger worker task to upload the pings in the background
             triggerWorkManager(context)
@@ -295,6 +302,11 @@ class MetricsPingSchedulerTest {
                     .getJSONObject("string")
                     .getString("telemetry.string_metric")
             )
+            assertEquals(
+                "The reason should be 'overdue'",
+                "overdue",
+                metricsJson.getJSONObject("ping_info").getString("reason")
+            )
         } finally {
             server.shutdown()
         }
@@ -311,9 +323,14 @@ class MetricsPingSchedulerTest {
         val mpsSpy =
             spy(MetricsPingScheduler(context))
         val overdueTestDate = "2015-07-05T12:36:00-06:00"
+
         mpsSpy.updateSentDate(overdueTestDate)
 
-        verify(mpsSpy, never()).collectPingAndReschedule(any(), eq(true))
+        verify(mpsSpy, never()).collectPingAndReschedule(
+            any(),
+            eq(true),
+            eq(Pings.metricsReasonCodes.overdue)
+        )
 
         // Make sure to return the fake date when requested.
         doReturn(fakeNow).`when`(mpsSpy).getCalendarInstance()
@@ -328,7 +345,11 @@ class MetricsPingSchedulerTest {
         assertEquals(fakeNow.time, mpsSpy.getLastCollectedDate())
 
         // Verify that we're immediately collecting.
-        verify(mpsSpy, times(1)).collectPingAndReschedule(fakeNow, true)
+        verify(mpsSpy, times(1)).collectPingAndReschedule(
+            fakeNow,
+            true,
+            Pings.metricsReasonCodes.overdue
+        )
     }
 
     @Test
@@ -343,13 +364,9 @@ class MetricsPingSchedulerTest {
         val mpsSpy =
             spy(MetricsPingScheduler(context))
 
-        // Inject the application version as already recorded, so we don't hit the case
-        // where the ping is sent due to a version change.
-        mpsSpy.isDifferentVersion()
-
         mpsSpy.updateSentDate(getISOTimeString(fakeNow, truncateTo = TimeUnit.Day))
 
-        verify(mpsSpy, never()).schedulePingCollection(any(), anyBoolean())
+        verify(mpsSpy, never()).schedulePingCollection(any(), anyBoolean(), eq(Pings.metricsReasonCodes.overdue))
 
         // Make sure to return the fake date when requested.
         doReturn(fakeNow).`when`(mpsSpy).getCalendarInstance()
@@ -358,9 +375,17 @@ class MetricsPingSchedulerTest {
         mpsSpy.schedule()
 
         // Verify that we're scheduling for the next day and not collecting immediately.
-        verify(mpsSpy, times(1)).schedulePingCollection(fakeNow, sendTheNextCalendarDay = true)
-        verify(mpsSpy, never()).schedulePingCollection(fakeNow, sendTheNextCalendarDay = false)
-        verify(mpsSpy, never()).collectPingAndReschedule(any(), eq(true))
+        verify(mpsSpy, times(1)).schedulePingCollection(
+            fakeNow,
+            sendTheNextCalendarDay = true,
+            reason = Pings.metricsReasonCodes.tomorrow
+        )
+        verify(mpsSpy, never()).schedulePingCollection(
+            eq(fakeNow),
+            sendTheNextCalendarDay = eq(false),
+            reason = any()
+        )
+        verify(mpsSpy, never()).collectPingAndReschedule(any(), eq(true), any())
     }
 
     @Test
@@ -375,10 +400,6 @@ class MetricsPingSchedulerTest {
         val mpsSpy =
             spy(MetricsPingScheduler(context))
 
-        // Inject the application version as already recorded, so we don't hit the case
-        // where the ping is sent due to a version change.
-        mpsSpy.isDifferentVersion()
-
         val fakeYesterday = Calendar.getInstance()
         fakeYesterday.time = fakeNow.time
         fakeYesterday.add(Calendar.DAY_OF_MONTH, -1)
@@ -387,15 +408,23 @@ class MetricsPingSchedulerTest {
         // Make sure to return the fake date when requested.
         doReturn(fakeNow).`when`(mpsSpy).getCalendarInstance()
 
-        verify(mpsSpy, never()).schedulePingCollection(any(), anyBoolean())
+        verify(mpsSpy, never()).schedulePingCollection(any(), anyBoolean(), any())
 
         // Trigger the startup check.
         mpsSpy.schedule()
 
         // Verify that we're scheduling for today, but not collecting immediately.
-        verify(mpsSpy, times(1)).schedulePingCollection(fakeNow, sendTheNextCalendarDay = false)
-        verify(mpsSpy, never()).schedulePingCollection(fakeNow, sendTheNextCalendarDay = true)
-        verify(mpsSpy, never()).collectPingAndReschedule(any(), eq(true))
+        verify(mpsSpy, times(1)).schedulePingCollection(
+            fakeNow,
+            sendTheNextCalendarDay = false,
+            reason = Pings.metricsReasonCodes.today
+        )
+        verify(mpsSpy, never()).schedulePingCollection(
+            eq(fakeNow),
+            sendTheNextCalendarDay = eq(true),
+            reason = any()
+        )
+        verify(mpsSpy, never()).collectPingAndReschedule(any(), eq(true), reason = any())
     }
 
     @Test
@@ -409,11 +438,10 @@ class MetricsPingSchedulerTest {
         val mpsSpy =
             spy(MetricsPingScheduler(context))
         mpsSpy.sharedPreferences.edit().clear().apply()
-        // Inject the application version as already recorded, so we don't hit the case
-        // where the ping is sent due to a version change.
+        // Restore the version number, so we don't get an "upgrade" reason ping
         mpsSpy.isDifferentVersion()
 
-        verify(mpsSpy, never()).collectPingAndReschedule(any(), anyBoolean())
+        verify(mpsSpy, never()).collectPingAndReschedule(any(), anyBoolean(), any())
 
         // Make sure to return the fake date when requested.
         doReturn(fakeNow).`when`(mpsSpy).getCalendarInstance()
@@ -422,8 +450,12 @@ class MetricsPingSchedulerTest {
         mpsSpy.schedule()
 
         // Verify that we're immediately collecting.
-        verify(mpsSpy, never()).collectPingAndReschedule(fakeNow, true)
-        verify(mpsSpy, times(1)).schedulePingCollection(fakeNow, sendTheNextCalendarDay = false)
+        verify(mpsSpy, never()).collectPingAndReschedule(eq(fakeNow), eq(true), any())
+        verify(mpsSpy, times(1)).schedulePingCollection(
+            fakeNow,
+            sendTheNextCalendarDay = false,
+            reason = Pings.metricsReasonCodes.today
+        )
     }
 
     @Test
@@ -440,7 +472,7 @@ class MetricsPingSchedulerTest {
         mpsSpy.schedule()
 
         // Verify that we're immediately collecting.
-        verify(mpsSpy, times(1)).collectPingAndReschedule(any(), anyBoolean())
+        verify(mpsSpy, times(1)).collectPingAndReschedule(any(), anyBoolean(), eq(Pings.metricsReasonCodes.upgrade))
     }
 
     @Test
@@ -525,8 +557,10 @@ class MetricsPingSchedulerTest {
         val mpsSpy =
             spy(MetricsPingScheduler(context))
         mpsSpy.sharedPreferences.edit().clear().apply()
+        // Restore the version number, so we don't get an "upgrade" reason ping
+        mpsSpy.isDifferentVersion()
 
-        verify(mpsSpy, never()).collectPingAndReschedule(any(), anyBoolean())
+        verify(mpsSpy, never()).collectPingAndReschedule(any(), anyBoolean(), any())
 
         // Make sure to return the fake date when requested.
         doReturn(fakeNow).`when`(mpsSpy).getCalendarInstance()
@@ -543,8 +577,8 @@ class MetricsPingSchedulerTest {
         )
 
         // Verify that we're immediately collecting.
-        verify(mpsSpy, times(1)).collectPingAndReschedule(fakeNow, true)
-        verify(mpsSpy, never()).schedulePingCollection(fakeNow, sendTheNextCalendarDay = false)
+        verify(mpsSpy, times(1)).collectPingAndReschedule(fakeNow, true, reason = Pings.metricsReasonCodes.overdue)
+        verify(mpsSpy, never()).schedulePingCollection(eq(fakeNow), sendTheNextCalendarDay = eq(false), reason = any())
     }
 
     @Test
@@ -557,10 +591,18 @@ class MetricsPingSchedulerTest {
         assertNull(Glean.metricsPingScheduler.timer)
 
         // Manually schedule a collection task for today.
-        Glean.metricsPingScheduler.schedulePingCollection(Calendar.getInstance(), sendTheNextCalendarDay = false)
+        Glean.metricsPingScheduler.schedulePingCollection(
+            Calendar.getInstance(),
+            sendTheNextCalendarDay = false,
+            reason = Pings.metricsReasonCodes.overdue
+        )
 
         // We expect the worker to be scheduled.
         assertNotNull(Glean.metricsPingScheduler.timer)
+
+        Glean.metricsPingScheduler.cancel()
+
+        resetGlean(clearStores = true)
     }
 
     @Test
@@ -568,7 +610,7 @@ class MetricsPingSchedulerTest {
         val context = ApplicationProvider.getApplicationContext<Context>()
         val mps = MetricsPingScheduler(context)
 
-        mps.schedulePingCollection(Calendar.getInstance(), true)
+        mps.schedulePingCollection(Calendar.getInstance(), true, reason = Pings.metricsReasonCodes.overdue)
 
         // Verify that the worker is enqueued
         assertNotNull("MetricsPingWorker is enqueued",