+++ glean-core/preview/examples/prototype.rs
@@ -11,7 +11,8 @@ use glean_preview as glean;
 use glean_preview::{metrics::PingType, ClientInfoMetrics, Configuration, Error};
 
 #[allow(non_upper_case_globals)]
-pub static PrototypePing: Lazy<PingType> = Lazy::new(|| PingType::new("prototype", true, true));
+pub static PrototypePing: Lazy<PingType> =
+    Lazy::new(|| PingType::new("prototype", true, true, vec![]));
 
 fn main() -> Result<(), Error> {
     env_logger::init();
@@ -42,7 +43,7 @@ fn main() -> Result<(), Error> {
     glean::initialize(cfg, client_info)?;
     glean::register_ping_type(&PrototypePing);
 
-    if glean::submit_ping_by_name("prototype") {
+    if glean::submit_ping_by_name("prototype", None) {
         log::info!("Successfully collected a prototype ping");
     } else {
         log::info!("Prototype ping failed");