+++ glean-core/src/ping/mod.rs
@@ -108,7 +108,7 @@ impl PingMaker {
         (start_time_data, end_time_data)
     }
 
-    fn get_ping_info(&self, glean: &Glean, storage_name: &str) -> JsonValue {
+    fn get_ping_info(&self, glean: &Glean, storage_name: &str, reason: Option<&str>) -> JsonValue {
         let (start_time, end_time) = self.get_start_end_times(glean, storage_name);
         let mut map = json!({
             "ping_type": storage_name,
@@ -116,6 +116,12 @@ impl PingMaker {
             "start_time": start_time,
             "end_time": end_time,
         });
+
+        if let Some(reason) = reason {
+            map.as_object_mut()
+                .unwrap() // safe unwrap, we created the object above
+                .insert("reason".to_string(), JsonValue::String(reason.to_string()));
+        };
 
         // Get the experiment data, if available.
         if let Some(experiment_data) =
@@ -162,12 +168,18 @@ impl PingMaker {
     ///
     /// * `glean` - the Glean instance to collect data from.
     /// * `ping` - the ping to collect for.
+    /// * `reason` - an optional reason code to include in the ping.
     ///
     /// ## Return value
     ///
     /// Returns a fully assembled JSON representation of the ping payload.
     /// If there is no data stored for the ping, `None` is returned.
-    pub fn collect(&self, glean: &Glean, ping: &PingType) -> Option<JsonValue> {
+    pub fn collect(
+        &self,
+        glean: &Glean,
+        ping: &PingType,
+        reason: Option<&str>,
+    ) -> Option<JsonValue> {
         info!("Collecting {}", ping.name);
 
         let metrics_data = StorageManager.snapshot_as_json(glean.storage(), &ping.name, true);
@@ -181,7 +193,7 @@ impl PingMaker {
             info!("Storage for {} empty. Ping will still be sent.", ping.name);
         }
 
-        let ping_info = self.get_ping_info(glean, &ping.name);
+        let ping_info = self.get_ping_info(glean, &ping.name, reason);
         let client_info = self.get_client_info(glean, ping.include_client_id);
 
         let mut json = json!({
@@ -206,13 +218,19 @@ impl PingMaker {
     ///
     /// * `glean` - the Glean instance to collect data from.
     /// * `ping` - the ping to collect for.
+    /// * `reason` - an optional reason code to include in the ping.
     ///
     /// ## Return value
     ///
     /// Returns a fully assembled ping payload in a string encoded as JSON.
     /// If there is no data stored for the ping, `None` is returned.
-    pub fn collect_string(&self, glean: &Glean, ping: &PingType) -> Option<String> {
-        self.collect(glean, ping)
+    pub fn collect_string(
+        &self,
+        glean: &Glean,
+        ping: &PingType,
+        reason: Option<&str>,
+    ) -> Option<String> {
+        self.collect(glean, ping, reason)
             .map(|ping| ::serde_json::to_string_pretty(&ping).unwrap())
     }